+++ rust/azure_iot_operations_protocol/src/common.rs
@@ -18,6 +18,9 @@ pub(crate) mod topic_processor;
 /// This module contains string values for Azure IoT Operations Protocol defined user properties.
 pub(crate) mod user_properties;
 
+/// This module contains a convenience for dispatching to a receiver based on an ID.
+pub mod dispatcher;
+
 /// This module contains the sending side cloud event implementation for the Azure IoT Operations Protocol.
 pub(crate) mod cloud_event;
 pub use cloud_event::CloudEventSubject;
