+++ rust/azure_iot_operations_services/src/lib.rs
@@ -34,5 +34,3 @@ pub mod leased_lock;
 pub mod schema_registry;
 #[cfg(feature = "state_store")]
 pub mod state_store;
-
-mod common;
